Organize code for site header
