+++ components/site-header/compact.js
@@ -1,0 +1,108 @@
+import { rem } from 'polished'
+import { useState } from 'react'
+import ReactDOM from 'react-dom'
+import styled, { css } from 'styled-components'
+import HamburgerMenu from './hamburger-menu'
+import Links from './links'
+import Logo from './logo'
+
+const NAV_ID = 'site-navigation-small'
+
+const NavOverlayRoot = styled.div`
+  position: fixed;
+  display: none;
+  width: 100%;
+  height: 0px;
+  padding: 32px 42px;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #000000;
+  color: #ffffff;
+  flex-direction: column;
+  align-items: center;
+  cursor: pointer;
+
+  ${(p) =>
+    p.isOpen &&
+    css`
+      display: flex;
+      height: 100%;
+    `}
+
+  @media only screen and (min-width: 576px) {
+    display: none;
+    height: 0px;
+  }
+`
+
+const FauxHeader = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+`
+
+const Nav = styled.nav`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  height: 100%;
+  width: 100%;
+
+  & a {
+    color: #fff;
+    text-decoration: none;
+    margin-top: ${rem('8px')};
+    font-size: ${rem('36px')};
+    line-height: ${rem('60px')};
+
+    &:first-of-type {
+      margin-top: 16px;
+    }
+  }
+`
+
+const NavOverlay = ({ id, isOpen, closeMenu, ...props }) => {
+  // Must check for process.browser; Without this, document is undefined because document is not available when nextjs renders this server side
+  if (process.browser) {
+    return ReactDOM.createPortal(
+      <NavOverlayRoot id={id} isOpen={isOpen} {...props}>
+        <FauxHeader>
+          <Logo />
+          <HamburgerMenu
+            isOpen={isOpen}
+            handleClick={closeMenu}
+            ariaControlsId={id}
+          />
+        </FauxHeader>
+        <Nav>
+          <Links />
+        </Nav>
+      </NavOverlayRoot>,
+      document.querySelector('body'),
+    )
+  }
+
+  return null
+}
+
+const Compact = () => {
+  const [isOpen, setIsOpen] = useState(false)
+  const openMenu = () => setIsOpen(!isOpen)
+  const closeMenu = () => setIsOpen(false)
+
+  return (
+    <>
+      <HamburgerMenu
+        isOpen={isOpen}
+        handleClick={openMenu}
+        ariaControlsId={NAV_ID}
+      />
+      <NavOverlay id={NAV_ID} isOpen={isOpen} closeMenu={closeMenu} />
+    </>
+  )
+}
+
+export default Compact
